+++ src/components/icons/SearchIcon.vue
